feat(sks): abstract video overlay components
